--- src/arduino-ide/SIGNALDuino/helpers.cpp
@@ -1,1 +1,0 @@
-../../helpers.cpp
