--- src/helpers.cpp
@@ -1,42 +1,0 @@
-#include "helpers.h"
-
-/* 
- * =================================================
- *  helper functions for smaller sketch size for uC
- * =================================================
- * 0 - 28806 size nano with debug
- */
-
-void MSG_PRINTtoHEX(uint8_t a) {
-/* 
- * this function is the alternative to 
- * sprintf(b, "%02x", xxx(i));
- * output via MSG_PRINT
- */
-  if(a < 16) {
-    MSG_PRINT(0);
-  }
-  MSG_PRINT(a , HEX);
-}
-
-
-void DBG_PRINTtoHEX(uint8_t b) {
-/* 
- * this function is the alternative to 
- * sprintf(b, "%02x", yyy(i));
- * output via DBG_PRINT
- */
-DBG_PRINT("TESTO");
-#ifdef DEBUG
-  if(b < 16) {
-    DBG_PRINT(0);
-  }
-  //DBG_PRINT(b , HEX);
-#endif
-}
-
-/* 
- * =====================
- *  END helper functions
- * =====================
- */
