+++ src/helpers.h
@@ -1,8 +1,45 @@
 #pragma once
 
-
 #include "Arduino.h"
 #include "output.h"
 
-void MSG_PRINTtoHEX(uint8_t a);
-void DBG_PRINTtoHEX(uint8_t b);
+/* 
+ * =================================================
+ *  helper functions for smaller sketch size for uC
+ * =================================================
+ * 0 - 28806 size nano with debug
+ */
+
+
+void MSG_PRINTtoHEX(uint8_t a) {
+/* 
+ * this function is the alternative to 
+ * sprintf(b, "%02x", xxx(i));
+ * output via MSG_PRINT
+ */
+  if(a < 16) {
+    MSG_PRINT(0);
+  }
+  MSG_PRINT(a , HEX);
+}
+
+void DBG_PRINTtoHEX(uint8_t b) {
+/* 
+ * this function is the alternative to 
+ * sprintf(b, "%02x", yyy(i));
+ * output via DBG_PRINT
+ */
+#ifdef DEBUG
+  if(b < 16) {
+    DBG_PRINT(0);
+  }
+  //DBG_PRINT(b , HEX);
+#endif
+}
+
+
+/* 
+ * =====================
+ *  END helper functions
+ * =====================
+ */
